use alloc_percpu correctly
This is a long bug of ktap, not noticed for a long time.

The argument of alloc_percpu should be a type, not sizeof

Without this change, the module refcount will be large after ktap tracing,
this is because module refcount is also a percpu variable, which overrided by ktapvm

And sometimes event.annotate output is crashed(Use percpu memory), this patch also fix it.