Fix the bug on parsing lines in KALLSYMS
'\t' and ' ' can both be used to delimit words in lines of the file
/proc/kallsyms, so strtok should take both as the delimiter parameter.

There is the '\t' between iter->name and iter->module_name
that may cause the error.

When iter->modele_name[0], the parser takes '%s\t[%s]' as the third
token, and treats it as function name, which may lead to unmatch.