Using a correct symbol name to look up its address for PPC64
For PPC64, the PPC64 ABI introduces function descriptor, it
is a three doubleword data structure. Its first doubleword
contains the address of the entry point of the function.

The value of the symbol with the same name as the function
is the address of function descriptor, Conversely, Symbol names
with a dot(.) prefix are real entry point addresses.

Signed-off-by: Yang Wei <w90p710@gmail.com>