Refactor ffi related function calls
Use 'ffi' as the keyword for ffi related functions, which are called in
ffi.* way

Supported API:
- ffi.cdef[[ long string ]]: c definition
- ffi.cdef("string"): c definition
- ffi.cnew("int[5]"): c new

For implementation, parser is hacked to treat ffi.* expression in some
special way. When it finds that the first token is 'ffi', and the
second one is '.', it is directed to parse it as ffi related function
calls.

Signed-off-by: Yicheng Qin <qycqycqycqycqyc@gmail.com>